Moved react dev project to docs folder
