+++ .circleci/config.yml
@@ -48,7 +48,7 @@ jobs:
             - run:
                   name: NPM Package
                   command: |
-                       cd dist && npm pack
+                      cd dist && npm pack
             # Save the dist folder for use in future jobs.
             - persist_to_workspace:
                   root: .
@@ -65,10 +65,10 @@ jobs:
         steps:
             - checkout
             - attach_workspace:
-                at: .
+                  at: .
             - codecov/upload:
-                file: './components/coverage/clover.xml'
-                token: 5e6faafe-eaf0-4d71-a931-22f279d3e446
+                  file: './components/coverage/clover.xml'
+                  token: 5e6faafe-eaf0-4d71-a931-22f279d3e446
 
     # Builds the Storybook app using @brightlayer-ui/react-components persisted in /dist folder.
     build_storybook:
@@ -80,7 +80,7 @@ jobs:
         steps:
             - checkout
             - attach_workspace:
-                at: .
+                  at: .
             - restore_cache:
                   keys:
                       - v2-dependencies-react-{{ checksum "demos/storybook/yarn.lock" }}
@@ -141,7 +141,7 @@ jobs:
         steps:
             - checkout
             - attach_workspace:
-                at: .
+                  at: .
             - run:
                   name: Initialize Submodule
                   command: |
@@ -191,11 +191,11 @@ jobs:
                   at: .
             - restore_cache:
                   keys:
-                      - v2-dependencies-react-{{ checksum "demos/react-dev/yarn.lock" }}
+                      - v2-dependencies-react-{{ checksum "docs/react-dev/yarn.lock" }}
             - run:
                   name: Install Dependencies
                   command: |
-                      cd demos/react-dev
+                      cd docs/react-dev
                       yarn install --frozen-lockfile
 
                       # Install the packed version of @brightlayer-ui/react-components
@@ -204,22 +204,22 @@ jobs:
             - save_cache:
                   name: Save Cache
                   paths:
-                      - demos/react-dev/node_modules
-                  key: v2-dependencies-react-{{ checksum "demos/react-dev/yarn.lock" }}
+                      - docs/react-dev/node_modules
+                  key: v2-dependencies-react-{{ checksum "docs/react-dev/yarn.lock" }}
             - run:
                   name: Prettier Check
                   command: |
-                      cd demos/react-dev
+                      cd docs/react-dev
                       yarn prettier:check
             - run:
                   name: Lint
                   command: |
-                      cd demos/react-dev
+                      cd docs/react-dev
                       yarn lint
             - run:
                   name: Build React Dev
                   command: |
-                      cd demos/react-dev
+                      cd docs/react-dev
                       yarn build
             - run:
                   command: cat /sys/fs/cgroup/memory/memory.max_usage_in_bytes
@@ -264,34 +264,34 @@ jobs:
                       git commit -m "Automated deployment to GitHub Pages: ${CIRCLE_SHA1}" --allow-empty
                       git push
     publish:
-      docker:
-        - image: cimg/node:14.19.0-browsers
-      steps:
-        - checkout
-        - attach_workspace:
-            at: .
-        - run:
-            name: Authenticate with registry
-            command: |
-              echo "//registry.npmjs.org/:_authToken=$NPM_TOKEN" > ~/.npmrc
-              echo "//registry.yarnpkg.com/:_authToken=$NPM_TOKEN" >> ~/.npmrc
-        - run:
-            name: Publish @brightlayer-ui/react-components
-            command:  |
-              yarn publish:package -b $CIRCLE_BRANCH
+        docker:
+            - image: cimg/node:14.19.0-browsers
+        steps:
+            - checkout
+            - attach_workspace:
+                  at: .
+            - run:
+                  name: Authenticate with registry
+                  command: |
+                      echo "//registry.npmjs.org/:_authToken=$NPM_TOKEN" > ~/.npmrc
+                      echo "//registry.yarnpkg.com/:_authToken=$NPM_TOKEN" >> ~/.npmrc
+            - run:
+                  name: Publish @brightlayer-ui/react-components
+                  command: |
+                      yarn publish:package -b $CIRCLE_BRANCH
 
     tag:
-      docker:
-        - image: cimg/node:14.19.0-browsers
-      steps:
-        - checkout
-        - attach_workspace:
-            at: .
-        - gh/setup
-        - run:
-            name: Tag @brightlayer-ui/react-components
-            command: |
-              yarn tag:package -b $CIRCLE_BRANCH -s -blui-react-components
+        docker:
+            - image: cimg/node:14.19.0-browsers
+        steps:
+            - checkout
+            - attach_workspace:
+                  at: .
+            - gh/setup
+            - run:
+                  name: Tag @brightlayer-ui/react-components
+                  command: |
+                      yarn tag:package -b $CIRCLE_BRANCH -s -blui-react-components
 
 workflows:
     version: 2
@@ -299,8 +299,8 @@ workflows:
         jobs:
             - build_library
             # - build_showcase:
-                #   requires:
-                #       - build_library
+            #   requires:
+            #       - build_library
             # - build_storybook:
             #       requires:
             #           - build_library
@@ -316,7 +316,7 @@ workflows:
                   requires:
                       - build_library
                   filters:
-                       branches:
+                      branches:
                           only:
                               - master
                               - dev
@@ -326,15 +326,14 @@ workflows:
                   filters:
                       branches:
                           only:
-                             - master
+                              - master
             - coverage_report:
-                   requires:
-                        - build_library
-                   filters:
-                        branches:
-                            only:
-                                - master
+                  requires:
+                      - build_library
+                  filters:
+                      branches:
+                          only:
+                              - master
     react_dev:
         jobs:
             - build_reactdev
-
