+++ docs/public/favicon.ico
