+++ docs/public/logo192.png
