+++ docs/public/logo512.png
