+++ docs/public/manifest.json
