+++ docs/src/pages/home.tsx
