+++ docs/src/pages/index.tsx
