+++ docs/src/redux/listItemComponent.tsx
